change in portfolio entry creation layout
So some fixes are made for the portfolio entries

- there was a placeholder added to the editor, which was only ok for
the dossier, but not the other use of whole page. It was corrected, so
it will change to the type of page it is looking for.
Also the placeholder was set in a strange way (directly via the html)
It is now setable in the quill api while importing quill to a page

- the save button layout with finalize checkbox was not dynamic with
the fileRole, but is now.

- a portfolio entry has now a title which is also shown in for the
docent and student (see and closes #214)