fix portfolio entry selection
this fixes the wall of text for the selection of portfolio entries.
also it adds a button to the portfolio page and also a add group entry
button to the portfolio page itself. With this, it is possible for the
user to create a group portfolio entry with a button click instead of a
visibility change.